Split endianness utilities from HeadSpin.h out into new hsEndian.h
